requirements.txt
numpy
jax
flax
matplotlib
pillow
distrax